refactor(host): lazy-load top-level views, add Groups to sidebar
Lazy-load all top-level route views (Home, Students, Posts, Groups) via
React.lazy for code-splitting, requiring each to switch to a default
export. Wrap the Routes in a Suspense boundary to support this.

Inline the registerRemotes/loadRemote/ErrorBoundary wiring directly
into PostsView and GroupsView instead of sharing it through
createParentsGatewayRemoteView.tsx — registerRemotes is safe to call
more than once for the same remote, so each view can independently
register `pg` without conflict, and having each view fully self-
contained is simpler than threading a shared factory.

Also add a Groups nav item to the sidebar, alongside Posts, since it
now has a real route and is otherwise unreachable from the UI.
