+++ apps/host/src/App.tsx
@@ -1,13 +1,15 @@
+import { lazy, Suspense } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router';
 
 import { AppSidebar } from '~/components/Sidebar';
 import { SidebarInset, SidebarProvider, SidebarTrigger } from '~/components/ui/sidebar';
 import { TooltipProvider } from '~/components/ui/tooltip';
-import { GroupsView } from '~/containers/GroupsView';
-import { HomeView } from '~/containers/HomeView';
 import { NotFoundView } from '~/containers/NotFoundView';
-import { PostsView } from '~/containers/PostsView';
-import { StudentsView } from '~/containers/StudentsView';
+
+const HomeView = lazy(() => import('~/containers/HomeView'));
+const StudentsView = lazy(() => import('~/containers/StudentsView'));
+const PostsView = lazy(() => import('~/containers/PostsView'));
+const GroupsView = lazy(() => import('~/containers/GroupsView'));
 
 export default function App() {
   return (
@@ -19,13 +21,15 @@ export default function App() {
             <header className="tw:flex tw:h-14 tw:items-center tw:px-4 tw:md:hidden">
               <SidebarTrigger />
             </header>
-            <Routes>
-              <Route path="/" element={<HomeView />} />
-              <Route path="/students/*" element={<StudentsView />} />
-              <Route path="/posts/*" element={<PostsView />} />
-              <Route path="/groups/*" element={<GroupsView />} />
-              <Route path="*" element={<NotFoundView />} />
-            </Routes>
+            <Suspense fallback={null}>
+              <Routes>
+                <Route path="/" element={<HomeView />} />
+                <Route path="/students/*" element={<StudentsView />} />
+                <Route path="/posts/*" element={<PostsView />} />
+                <Route path="/groups/*" element={<GroupsView />} />
+                <Route path="*" element={<NotFoundView />} />
+              </Routes>
+            </Suspense>
           </SidebarInset>
         </SidebarProvider>
       </TooltipProvider>
