+++ apps/host/src/containers/StudentsView.tsx
@@ -1,4 +1,4 @@
-export function StudentsView() {
+export default function StudentsView() {
   return (
     <div className="tw:flex tw:flex-1 tw:items-center tw:justify-center tw:p-8">
       <h1 className="tw:text-2xl tw:font-semibold tw:text-muted-foreground">Student Insight</h1>
